fix(jest-config): rename src/boostrap/ to src/bootstrap/ and update all imports
Fixes a typo in the jest-config package directory name and updates the 43
test files in framework/core that imported from the misspelled path.

Fixes #4374
