+++ framework/core/js/tests/integration/admin/components/MailPage.test.ts
@@ -1,4 +1,4 @@
-import bootstrapAdmin from '@flarum/jest-config/src/boostrap/admin';
+import bootstrapAdmin from '@flarum/jest-config/src/bootstrap/admin';
 import MailPage from '../../../../src/admin/components/MailPage';
 import { app } from '../../../../src/admin';
 import mq from 'mithril-query';
