+++ framework/core/js/tests/integration/common/AlertManager.test.ts
@@ -1,4 +1,4 @@
-import bootstrapForum from '@flarum/jest-config/src/boostrap/forum';
+import bootstrapForum from '@flarum/jest-config/src/bootstrap/forum';
 import { app } from '../../../src/forum';
 import mq from 'mithril-query';
 import AlertManager from '../../../src/common/components/AlertManager';
