+++ framework/core/js/tests/integration/common/components/Badge.test.ts
@@ -1,4 +1,4 @@
-import bootstrapForum from '@flarum/jest-config/src/boostrap/forum';
+import bootstrapForum from '@flarum/jest-config/src/bootstrap/forum';
 import Badge from '../../../../src/common/components/Badge';
 import m from 'mithril';
 import mq from 'mithril-query';
