+++ framework/core/js/tests/integration/common/components/FormModal.test.ts
@@ -1,4 +1,4 @@
-import bootstrapForum from '@flarum/jest-config/src/boostrap/forum';
+import bootstrapForum from '@flarum/jest-config/src/bootstrap/forum';
 import FormModal from '../../../../src/common/components/FormModal';
 import ModalManagerState from '../../../../src/common/states/ModalManagerState';
 import m from 'mithril';
