+++ framework/core/js/tests/integration/common/components/Navigation.test.ts
@@ -1,4 +1,4 @@
-import bootstrapForum from '@flarum/jest-config/src/boostrap/forum';
+import bootstrapForum from '@flarum/jest-config/src/bootstrap/forum';
 import Navigation from '../../../../src/common/components/Navigation';
 import { app } from '../../../../src/forum';
 import mq from 'mithril-query';
