+++ js-packages/jest-config/src/bootstrap/admin.js
