+++ js-packages/jest-config/src/bootstrap/common.js
